% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acs_grade_lookup}
\alias{acs_grade_lookup}
\title{School grade lookup table}
\format{
A data frame with 18 rows and 3 variables:
\describe{
  \item{name}{acs variable name}
  \item{label}{acs variable label}
  \item{grade}{two-character grade classifier found in pums SCHG variable}
  ...
}
}
\usage{
acs_grade_lookup
}
\description{
Dataset to relate acs variables from B14007_ table (detailed school grade) to pums school grades
}
\keyword{datasets}
