% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acs_hhincome_lookup}
\alias{acs_hhincome_lookup}
\title{Household income lookup table}
\format{
A data frame with 18 rows and 3 variables:
\describe{
  \item{name}{acs variable name}
  \item{label}{acs variable label}
  \item{concept}{variable description}
  \item{income_min}{minimum income captured in variable extracted from label}
  \item{income_max}{maximum income captured in variable extracted from label}
  ...
}
}
\usage{
acs_hhincome_lookup
}
\description{
Dataset to relate acs variables from S1901_C01 tables (household income groups) to aid merging to pums income categories
}
\keyword{datasets}
