% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acs_hhsize_type_lookup}
\alias{acs_hhsize_type_lookup}
\title{Household size and type lookup table}
\format{
A data frame with 13 rows and 3 variables:
\describe{
  \item{name}{acs variable name}
  \item{hh_size}{number of persons living in household}
}
}
\usage{
acs_hhsize_type_lookup
}
\description{
Dataset to relate acs variables from B11016 tables (household size and type) household size and type variables in PUMS
}
\keyword{datasets}
