% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acs_occp_lookup}
\alias{acs_occp_lookup}
\title{Occupations codes}
\format{
A data frame with 563 rows and 6 variables:
\describe{
  \item{Code}{Specific codes corresponding to occupation}
  \item{Code2}{Occupation group}
  \item{Description}{Occupation description}
  \item{d2_code}{Two digit occupation group}
  \item{occp_match}{field used to merge pums and acs datasets. See `Get_Helpers.R`}
  \item{acs_var}{acs variable code corresponding to number of individuals in that occupation in the census geography}
  ...
}
}
\source{
\url{https://www2.census.gov/programs-surveys/acs/tech_docs/pums/code_lists/ACSPUMS2015_2019CodeLists.xlsx}
}
\usage{
acs_occp_lookup
}
\description{
Dataset to relate occupational codes to acs variables
}
\keyword{datasets}
