% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acs_race_eth_lookup}
\alias{acs_race_eth_lookup}
\title{Race and ethnicity lookup table}
\format{
A data frame with 18 rows and 3 variables:
\describe{
  \item{name}{acs variable name}
  \item{hispanic}{binary of hispanic or latino (1) or not hispanic or latino (0)}
  \item{race}{character description of race}
  ...
}
}
\usage{
acs_race_eth_lookup
}
\description{
Dataset to relate acs variables from B03002 tables (race and ethnicity totals) to race and ethnicity categories
}
\keyword{datasets}
