% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acs_sex_by_age_lookup}
\alias{acs_sex_by_age_lookup}
\title{Age by sex lookup table}
\format{
A data frame with 298 rows and 6 variables:
\describe{
  \item{name}{acs variable name}
  \item{label}{acs variable label}
  \item{concept}{variable description}
  \item{age_min}{minimum age captured in variable extracted from label}
  \item{age_max}{maximum age captured in variable extracted from label}
  \item{sex}{sex captured in variable extracted from label}
  \item{sex_num}{numeric sex where Male=1, female=2}
  ...
}
}
\usage{
acs_sex_by_age_lookup
}
\description{
Dataset to relate acs variables from B01001 tables (age by sex and age by sex by race/ethnicity) to age categories and sex
}
\keyword{datasets}
