% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Functions.R
\name{gen_ct_pop}
\alias{gen_ct_pop}
\title{Create synthetic population for target census tract}
\usage{
gen_ct_pop(CT, PTGT, HTGT, PSEED, HSEED)
}
\arguments{
\item{CT}{character containing name of CT as it appears in ACS data}

\item{PTGT}{list generated from `p_target` with person characteristics}

\item{HTGT}{list generated from `h_target` with household characteristics}

\item{PSEED}{data frame containing person pums data for target geography}

\item{HSEED}{data frame containing household pums data for target geography}
}
\value{
data.frame of all generated agents
}
\description{
Creates synthetic population from input target and seed data for a
}
\details{

}
