% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Functions.R
\name{gen_pop}
\alias{gen_pop}
\title{Create synthetic population}
\usage{
gen_pop(CTs)
}
\arguments{
\item{CT}{character containing name of CT as it appears in ACS data}

\item{PTGT}{list generated from `p_target` with person characteristics}

\item{HTGT}{list generated from `h_target` with household characteristics}

\item{PSEED}{data frame containing person pums data for SF county}

\item{HSEED}{data frame containing household pums data for SF county}
}
\value{
data.frame of all generated agents
}
\description{
Creates synthetic population from input target and seed data for a geography larger than a census tract
}
\details{

}
