% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gen_Pop.R
\name{gen_synth_pop}
\alias{gen_synth_pop}
\title{Generate population}
\usage{
gen_synth_pop(FIPS, CHARS, OUT_DIR, KEEP_ALL = TRUE)
}
\arguments{
\item{FIPS}{US FIPS code for corresponding geography. If length 2, will generate for corresponding state. If length 5, will generate for corresponding county}

\item{CHARS}{Agent characteristics to include. See `details`}

\item{OUT_DIR}{Directory to save generated population to. Also will house intermediate downloaded files if KEEP_ALL is TRUE}

\item{KEEP_ALL}{Logical of whether to keep all intermediate files used to generate populaiton. Defaults to TRUE}
}
\value{

}
\description{
One-liner function to generate population
}
\details{

}
