% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Data.R
\name{get_pums}
\alias{get_pums}
\title{Get PUMS}
\usage{
get_pums(YEAR, FIVE_YEAR = TRUE, STATE, LEVEL, KEEP = FALSE, KEEP_DIR)
}
\arguments{
\item{YEAR}{survey year to download data.}

\item{FIVE_YEAR}{whether to download five year survey data or one year. defaults to five year, set to FALSE for 1 year}

\item{STATE}{two-letter state abbreviation for state to grab data for}

\item{LEVEL}{person or household level data download. "p" for person, "h" for household}

\item{KEEP}{logical of whether to keep raw downloaded zip file, defaults to FALSE}

\item{KEEP_DIR}{directory at which to store downloaded zip file}
}
\value{

}
\description{
Interfaces with ACS PUMS ftp site to download pums data for specified state
}
\details{

}
