% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Functions.R
\name{h_seed}
\alias{h_seed}
\title{Household seed}
\usage{
h_seed()
}
\arguments{
\item{h_pums}{household PUMS data (e.g. downloaded from `get_pums`)}

\item{OCCP}{data table/tibble that contains occupational data for target geography}

\item{AGE_SEX}{data table/tibble that contains ACS data for target geography stratified by sex and age}

\item{RACE}{data table/tibble that contains ACS data for target geography stratified by race/ethnicity}
}
\value{
list containing targets for occupation, age, sex and race for the CT
}
\description{
Creates household seed for input to ipf
}
\details{

}
