% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Functions.R
\name{h_target}
\alias{h_target}
\title{Household target}
\usage{
h_target(CT, INCOME, SIZE)
}
\arguments{
\item{CT}{character containing name of CT as it appears in ACS data}

\item{INCOME}{data table/tibble that contains household income data for target geography all income is in 2018-inflation-adjusted dollars}

\item{SIZE}{data table/tibble that contains ACS data for target geography stratified by household size}
}
\value{
list containing targets for income and size distribution of households within the CT
}
\description{
Creates household target for input to ipf
}
\details{

}
