% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OCC_list}
\alias{OCC_list}
\title{Occupations codes}
\format{
A data frame with 74091 rows and 4 variables:
\describe{
  \item{Code}{Specific codes corresponding to occupation}
  \item{Code2}{Occupation group}
  \item{Description}{Occupation description}
  \item{d2_code}{Two digit occupation group}
  \item{New_Code}{Code used to condense into groups}
  ...
}
}
\source{
\url{https://www2.census.gov/programs-surveys/acs/tech_docs/pums/code_lists/ACSPUMS2014_2018CodeLists.xlsx}
}
\usage{
OCC_list
}
\description{
Dataset to condense occupation codes into groups
}
\keyword{datasets}
