% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_Functions.R
\name{p_seed}
\alias{p_seed}
\title{Person seed}
\usage{
p_seed()
}
\arguments{
\item{CT}{character containing name of CT as it appears in ACS data}

\item{OCCP}{data table/tibble that contains occupational data for target geography}

\item{AGE_SEX}{data table/tibble that contains ACS data for target geography stratified by sex and age}

\item{RACE}{data table/tibble that contains ACS data for target geography stratified by race/ethnicity}
}
\value{
list containing targets for occupation, age, sex and race for the CT
}
\description{
Creates person seed for input to ipf
}
\details{

}
