% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pums_to_acs}
\alias{pums_to_acs}
\title{Variables of interest relation to PUMS variable names and acs data table codes}
\format{
A list of variables where each list contains a PUMS code or codes and an ACS data table code or codes which is past to `tidycensus` functions for downloading data from user input attributes. See `R/rsp_get_data()` and `Analysis/ACS_PUMS_List.R`
}
\usage{
pums_to_acs
}
\description{
`pums_to_acs <- readRDS(here::here(data/pums_to_acs.rds))`
}
\keyword{datasets}
