% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSP_process_functions.R
\name{rsp_clean_gq_pop}
\alias{rsp_clean_gq_pop}
\title{Clean group quarters synthetic population to match characteristics in synth pop}
\usage{
rsp_clean_gq_pop(
  gq_pop,
  chars,
  fips_use,
  p_age_breaks = c(0, 10, 20, 30, 40, 50, 60, 70, 80, 150)
)
}
\arguments{
\item{gq_pop}{Synthesized group quarters population, e.g. from `rsp_gq_synth()`}

\item{chars}{person-level characteristics to be synthesized. Will be used to aggregate and summarize data in `gq_pop`}

\item{fips_use}{fips code corresponding to target area to synthesize group quarters population}

\item{p_age_breaks}{vector of breaks at which to divide age categories, defaults to deciles}
}
\value{
list with names as in chars with marginal totals of each characteristic represented in gq population.
}
\description{
Used to clean column names and discretized variables to match synthetic population
}
\details{

}
