% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSP_get_functions.R
\name{rsp_get_acs}
\alias{rsp_get_acs}
\title{Get pums data}
\usage{
rsp_get_acs(VARS, SURVEY = "acs5", STATES, LEVEL, YEAR)
}
\arguments{
\item{VARS}{character vector containing desired characteristics of agents. See details for currently supported characteristics.}

\item{SURVEY}{Survey type to get data from. "acs1" or "acs5"}

\item{STATES}{two letter state abbreviation or character vector containing multiple}

\item{LEVEL}{Level to return data for, options are "State", "County", "Tract"}

\item{YEAR}{year from which data is desired}
}
\value{

}
\description{
Get ACS data for desired variables
}
\details{
See included data file `pums_to_acs` for currently supported characteristics and Tutorial/Add_New_Characteristics for guide on adding additional characteristics
}
