% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSP_get_functions.R
\name{rsp_get_gq}
\alias{rsp_get_gq}
\title{Get group quarters target data}
\usage{
rsp_get_gq(STATES, LEVEL)
}
\arguments{
\item{STATES}{Single state or character vector of states (two letter abbreviations) to download data}

\item{LEVEL}{Level to return data for, options are "State", "County", "Tract"}
}
\value{
list containing data frame of group quarters data for each state at the indicated level.
}
\description{
Interfaces with census api to download census summary file f1 and extract group quarters data. Currently fetches data from the 2010 census, will be updated for 2020 census when data is made available
}
\details{
See <https://gl-li.netlify.app/2017/08/29/process-2010-census-data-with-data-table/> for useful information on processing raw census supplement data. Each returned data frame in the list of states will contain data frame with id record from geography census file, the state, county, and tract fips code (as relevant per input LEVEL), and then 63 columns corresponding to variables in table P43 (see https://www.census.gov/prod/cen2010/doc/sf1.pdf page 6-55)
}
