% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSP_synth_functions.R
\name{rsp_gq_synth}
\alias{rsp_gq_synth}
\title{Synthesize group quarters}
\usage{
rsp_gq_synth(fips_use, gq_seed, gq_tgt)
}
\arguments{
\item{fips_use}{fips code corresponding to target area to synthesize group quarters population}

\item{gq_seed}{seed data frame for group quarters population}

\item{gq_tgt}{target list of data frames with names corresponding to characteristics in columns of seed for group wuarters population}
}
\value{
data frame of synthesized population with all characteristics contained in `gq_seed` maintained
}
\description{

}
\details{

}
