% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSP_synth_functions.R
\name{rsp_hhp_synth}
\alias{rsp_hhp_synth}
\title{Synthesize population}
\usage{
rsp_hhp_synth(fips_use, hh_seed, p_seed, hh_tgt, p_tgt, gq_pop = NULL)
}
\arguments{
\item{fips_use}{fips code corresponding to target area to synthesize group quarters population}

\item{hh_seed}{seed data frame for households}

\item{p_seed}{seed data frame for persons}

\item{hh_tgt}{target list of data frames with names corresponding to characteristics in columns of seed for households}

\item{p_tgt}{target list of data frames with names corresponding to characteristics in columns of seed for persons}

\item{gq_pop}{synthesized group quarters population. if included, characteristics will be summarised and subtracted from the marginal (target) person totals of all characteristics}
}
\value{
data frame of synthesized population with all characteristics contained in `gq_seed` maintained
}
\description{

}
\details{

}
