% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSP_process_functions.R
\name{rsp_process_gq_seed}
\alias{rsp_process_gq_seed}
\title{Process group quarters seed data from pums}
\usage{
rsp_process_gq_seed(pums_dat, pumas)
}
\arguments{
\item{pums_dat}{pums dataset, e.g. from `rsp_get_pums()`}

\item{pumas}{puma areas to include}
}
\value{
data frame containing person pums records for individuals in group quarters
}
\description{
Processes pums data to extract all those in group quarters and generate variables to match information in census table P43 that serves as target data
}
\details{

}
