% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSP_process_functions.R
\name{rsp_process_gq_tgt}
\alias{rsp_process_gq_tgt}
\title{Process group quarters target data from acs supplement file 1}
\usage{
rsp_process_gq_tgt(gq_dat, fips_use)
}
\arguments{
\item{gq_dat}{data frame returned as element of list from `rsp_get_gq()` with processed data from census table P43}

\item{fips_use}{fips code to restrict acs data by.}
}
\value{
list containing named elements that match columns names of group quarters variables in seed data. in each list is a data frames where columns match categories of the variables in seed data.
}
\description{

}
\details{

}
