% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSP_process_functions.R
\name{rsp_process_hh_seed}
\alias{rsp_process_hh_seed}
\title{Process household level seed data from pums}
\usage{
rsp_process_hh_seed(
  pums_dat,
  pumas,
  hh_vars,
  hh_income_breaks = c(-100, 50, 100, Inf) * 1000
)
}
\arguments{
\item{pums_dat}{pums dataset, e.g. from `rsp_get_pums`}

\item{pumas}{puma areas to include}

\item{hh_vars}{vector of household variable names found in pums dat to process}

\item{hh_income_breaks}{vector of breaks at which to divide hh income categories}
}
\value{
data frame containing household pums records with all household level characteristics of interest
}
\description{

}
\details{

}
