% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSP_process_functions.R
\name{rsp_process_hh_tgt}
\alias{rsp_process_hh_tgt}
\title{Process household level target data from acs totals}
\usage{
rsp_process_hh_tgt(
  acs_hh_dat,
  chars,
  fips_use,
  hh_income_breaks = c(-100, 50, 100, 1e+05) * 1000
)
}
\arguments{
\item{acs_hh_dat}{list of data frames containing household level data from acs/census download, filtered to contain household variables}

\item{chars}{vector of characteristics to include in population. will only look for data in `acs_hh_dat` that is indicated by characteristics in this vector. e.g. will look for and process household income data if "HH_Income" is an element of `chars`}

\item{fips_use}{fips code to restrict acs data by.}

\item{hh_income_breaks}{vector of breaks at which to divide hh income categories. be sure to match this to same breaks used in `rsp_process_hh_seed()` to ensure levels match}
}
\value{
list containing named elements that match columns names of household level variables in seed data. in each list is a data frames where columns match categories of the variables in seed data.
}
\description{

}
\details{

}
