% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSP_process_functions.R
\name{rsp_process_p_seed}
\alias{rsp_process_p_seed}
\title{Process person level seed data from pums}
\usage{
rsp_process_p_seed(
  pums_dat,
  pumas,
  p_vars,
  p_age_breaks = c(0, 10, 20, 30, 40, 50, 60, 70, 80, 150)
)
}
\arguments{
\item{pums_dat}{pums dataset, e.g. from `rsp_get_pums`}

\item{pumas}{puma areas to include}

\item{p_vars}{vector of person level  variable names found in pums_dat to process}

\item{p_age_breaks}{vector of breaks at which to divide age categories, defaults to deciles}
}
\value{
data frame containing person-level pums records with all household level characteristics of interest
}
\description{

}
\details{

}
