% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSP_process_functions.R
\name{rsp_process_p_tgt}
\alias{rsp_process_p_tgt}
\title{Process person level target data from acs totals}
\usage{
rsp_process_p_tgt(
  acs_p_dat,
  chars,
  fips_use,
  p_age_breaks = c(0, 10, 20, 30, 40, 50, 60, 70, 80, 150)
)
}
\arguments{
\item{acs_p_dat}{list of data frames containing person level data from acs/census download, filtered to contain person variables}

\item{chars}{vector of characteristics to include in population. will only look for data in `acs_p_dat` that is indicated by characteristics in this vector. e.g. will look for and process age data if "Age" is an element of `chars`}

\item{fips_use}{fips code to restrict acs data by.}

\item{p_age_breaks}{vector of breaks at which to divide hh income categories. be sure to match this to same breaks used in `rsp_process_hh_seed()` to ensure levels match}
}
\value{
list containing named elements that match columns names of household level variables in seed data. in each list is a data frames where columns match categories of the variables in seed data.
}
\description{

}
\details{

}
